@@ -1,52 +1,55 @@
 {
-    "name": "andrii-hrechyn/auto-documentation",
-    "description": "Auto generated documentation for laravel",
-    "type": "library",
-    "license": "MIT",
-    "authors": [
-        {
-            "role": "Developer",
-            "name": "Andriy Hrechyn",
-            "email": "andriy.hrechyn@gmail.com"
-        }
-    ],
-    "autoload": {
-        "psr-4": {
-            "AutoDocumentation\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "AutoDocumentation\\Tests\\": "tests/"
-        }
-    },
-    "require": {
-        "php": "^8.1",
-        "illuminate/support": "^7.0|^8.0|^9.0|^10.0",
-        "symfony/process": "^6.2",
-        "symfony/yaml": "^6.1"
-    },
-    "require-dev": {
-        "orchestra/testbench": "^8.0",
-        "pestphp/pest": "^2.0",
-        "pestphp/pest-plugin-laravel": "^2.0",
-        "pestphp/pest-plugin-mock": "^v2.0.0"
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "config": {
-        "sort-packages": true,
-        "preferred-install": "dist",
-        "optimize-autoloader": true,
-        "allow-plugins": {
-            "pestphp/pest-plugin": true
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "AutoDocumentation\\AutoDocumentationServiceProvider"
-            ]
-        }
+  "name": "andrii-hrechyn/auto-documentation",
+  "description": "Auto generated documentation for laravel",
+  "type": "library",
+  "license": "MIT",
+  "authors": [
+    {
+      "role": "Developer",
+      "name": "Andriy Hrechyn",
+      "email": "andriy.hrechyn@gmail.com"
     }
+  ],
+  "autoload": {
+    "psr-4": {
+      "AutoDocumentation\\": "src/"
+    },
+    "files": [
+      "src/helpers.php"
+    ]
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "AutoDocumentation\\Tests\\": "tests/"
+    }
+  },
+  "require": {
+    "php": "^8.3",
+    "illuminate/support": "^7.0|^8.0|^9.0|^10.0|^11.0",
+    "symfony/process": "^6.2",
+    "symfony/yaml": "^6.2"
+  },
+  "require-dev": {
+    "orchestra/testbench": "^8.0",
+    "pestphp/pest": "^2.0",
+    "pestphp/pest-plugin-laravel": "^2.0",
+    "pestphp/pest-plugin-mock": "^v2.0.0"
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "config": {
+    "sort-packages": true,
+    "preferred-install": "dist",
+    "optimize-autoloader": true,
+    "allow-plugins": {
+      "pestphp/pest-plugin": true
+    }
+  },
+  "extra": {
+    "laravel": {
+      "providers": [
+        "AutoDocumentation\\AutoDocumentationServiceProvider"
+      ]
+    }
+  }
 }
