@@ -3,18 +3,24 @@
 namespace AutoDocumentation\Components;
 
 use AutoDocumentation\Enums\ComponentType;
-use AutoDocumentation\Parameters\Parameter;
+use AutoDocumentation\Enums\ParameterIn;
+use AutoDocumentation\Parameter;
 
-abstract class ParameterComponent extends Component
+abstract class ParameterComponent extends SimpleComponent
 {
     public function type(): ComponentType
     {
-        return ComponentType::parameters;
+        return ComponentType::PARAMETER;
     }
 
-    public function name(): string
+    public function getName(): string
     {
-        return $this->content()->name;
+        return $this->content()->getName();
+    }
+
+    protected function parameter(string $name, ParameterIn $in): Parameter
+    {
+        return Parameter::make($name, $in);
     }
 
     abstract public function content(): Parameter;
