@@ -26,9 +26,50 @@ class GenerateDocumentation extends Command
      */
     public function handle(Documentation $documentation)
     {
+        $doc = new \Docs\Documentation();
+
+        $api = $doc->resolveOpenApi();
+
+
+//        $api = (new OpenApi())
+//            ->servers([
+//                Server::make('http://localhost:8000/{locale}')
+//                    ->description('My description')
+//                    ->variable(Variable::make('locale', 'en')->enum(['en', 'ua'])),
+//            ])
+//            ->security(SanctumAuth::make())
+//            ->paths([
+//                Path::make('/users')
+//                    ->methods([
+//                        Method::make('get')
+//                            ->security(SanctumAuth::make())
+//                            ->parameter(UserIdParameter::make())
+//                            ->servers([
+//                                Server::make('http://bebebe:8000/{locale}')
+//                                    ->description('My description')
+//                                    ->variable(Variable::make('locale', 'en')->enum(['en', 'ua'])),
+//                            ])
+//                            ->request(
+//                                Request::make()
+//                                    ->content([
+//                                        Content::make('application/json')
+//                                            ->schema(
+//                                                TestComponent::make()
+//                                            ),
+//                                    ])
+//                            )
+//                            ->responses([
+//                                SuccessfulResponse::make(),
+//                                NoContentResponse::make(),
+//                            ]),
+//                    ]),
+//            ]);
+
+//        dd($api->generate());
+
         $this->warn('Generating documentation ...');
 
-        $documentation->generate();
+        $documentation->generate($api);
 
         $this->info('Documentation generated! Visit: '.route('auto-documentation.documentation'));
     }
