@@ -8,27 +8,17 @@ use Symfony\Component\Yaml\Yaml;
 class Documentation
 {
     protected Filesystem $filesystem;
+    protected Normalizer $normalizer;
 
-    public function __construct(Filesystem $filesystem)
+    public function __construct(Filesystem $filesystem, Normalizer $normalizer)
     {
         $this->filesystem = $filesystem;
+        $this->normalizer = $normalizer;
     }
 
-    public function generate(): void
+    public function generate($a): void
     {
-        $this->loadPaths(config('auto-documentation.paths.source'));
-
-        $this->store(OpenApi::instance(), config('auto-documentation.paths.generated-doc').'/documentation.yaml');
-    }
-
-    public function loadPaths(string $documentationPath)
-    {
-        (new Loader($this->filesystem))->load($documentationPath);
-    }
-
-    public function generateYaml(OpenApi $api): string
-    {
-        return Yaml::dump($api->generate(), 25, 2, Yaml::DUMP_EMPTY_ARRAY_AS_SEQUENCE);
+        $this->store($a, config('auto-documentation.paths.generated-doc').'/documentation.yaml');
     }
 
     public function store(OpenApi $api, string $path): void
@@ -36,5 +26,10 @@ class Documentation
         $this->filesystem->ensureDirectoryExists(dirname($path));
 
         $this->filesystem->put($path, $this->generateYaml($api));
+    }
+
+    public function generateYaml(OpenApi $api): string
+    {
+        return Yaml::dump($this->normalizer->normalize($api->toArray()), 25, 2, Yaml::DUMP_EMPTY_ARRAY_AS_SEQUENCE);
     }
 }
