@@ -2,30 +2,17 @@
 
 namespace AutoDocumentation\Headers;
 
+use AutoDocumentation\Traits\HasDeprecated;
+use AutoDocumentation\Traits\HasDescription;
+use AutoDocumentation\Traits\HasExtensions;
+use AutoDocumentation\Traits\HasName;
+use AutoDocumentation\Traits\HasRequired;
+
 class Header
 {
-    protected ?string $description = null;
-    protected bool $required = false;
-    protected bool $deprecated = false;
-
-    public function description(string $description): self
-    {
-        $this->description = $description;
-
-        return $this;
-    }
-
-    public function required(): self
-    {
-        $this->required = true;
-
-        return $this;
-    }
-
-    public function deprecated(): self
-    {
-        $this->deprecated = true;
-
-        return $this;
-    }
+    use HasName;
+    use HasDescription;
+    use HasRequired;
+    use HasDeprecated;
+    use HasExtensions;
 }
