@@ -2,62 +2,99 @@
 
 namespace AutoDocumentation;
 
-class Info
+use AutoDocumentation\Traits\HasDescription;
+use AutoDocumentation\Traits\HasExtensions;
+use AutoDocumentation\Traits\HasTitle;
+use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Support\Str;
+
+class Info implements Arrayable
 {
-    protected string $description = '';
+    use HasTitle;
+    use HasDescription;
+    use HasExtensions;
+
+    protected string $version;
+
     protected ?string $termsOfService = null;
 
     protected ?Contact $contact = null;
+
     protected ?License $license = null;
 
-    public static function make(string $title, string $version): self
+    public function __construct(string $title, string $version)
     {
-        return OpenApi::instance()->info(new self($title, $version));
+        $this->title($title)
+            ->version($version);
     }
 
-    private function __construct(
-        protected readonly string $title,
-        protected readonly string $version
-    ) {
+    public static function make(string $title, string $version): static
+    {
+        return new static($title, $version);
     }
 
-    public function description(string $description): self
+    public function version(string $version): static
     {
-        $this->description = $description;
+        $this->version = $version;
 
         return $this;
     }
 
-    public function termsOfService(string $url): self
+    public function getVersion(): string
     {
-        $this->termsOfService = $url;
+        return $this->version;
+    }
+
+    public function termsOfService(string $termsOfService): static
+    {
+        if (!Str::isUrl($termsOfService)) {
+            //todo throw exception
+        }
+
+        $this->termsOfService = $termsOfService;
 
         return $this;
     }
 
-    public function contact(Contact $contact): self
+    public function getTermsOfService(): ?string
+    {
+        return $this->termsOfService;
+    }
+
+    public function contact(Contact $contact): static
     {
         $this->contact = $contact;
 
         return $this;
     }
 
-    public function license(License $license): self
+    public function getContact(): ?Contact
+    {
+        return $this->contact;
+    }
+
+    public function license(License $license): static
     {
         $this->license = $license;
 
         return $this;
     }
 
+    public function getLicense(): ?License
+    {
+        return $this->license;
+    }
+
     public function toArray(): array
     {
-        return [
-            'title'          => $this->title,
-            'description'    => $this->description,
-            'termsOfService' => $this->termsOfService,
-            'contact'        => $this->contact?->toArray(),
-            'license'        => $this->license?->toArray(),
-            'version'        => $this->version,
-        ];
+        return array_filter([
+            'title'          => $this->getTitle(),
+            'description'    => $this->getDescription(),
+            'termsOfService' => $this->getTermsOfService(),
+            'contact'        => $this->getContact()?->toArray(),
+            'license'        => $this->getLicense()?->toArray(),
+            'version'        => $this->getVersion(),
+            ...$this->getExtensions(),
+        ], fn ($value) => $value !== null);
     }
 }
