@@ -2,19 +2,33 @@
 
 namespace AutoDocumentation;
 
-class License
+use AutoDocumentation\Traits\HasExtensions;
+use AutoDocumentation\Traits\HasName;
+use AutoDocumentation\Traits\HasUrl;
+use Illuminate\Contracts\Support\Arrayable;
+
+class License implements Arrayable
 {
-    public function __construct(
-        protected readonly string $name,
-        protected readonly ?string $url = null
-    ) {
+    use HasName;
+    use HasUrl;
+    use HasExtensions;
+
+    public function __construct(string $name)
+    {
+        $this->name($name);
+    }
+
+    public static function make(string $name): static
+    {
+        return new static($name);
     }
 
     public function toArray(): array
     {
-        return [
-            'name' => $this->name,
-            'url'  => $this->url,
-        ];
+        return array_filter([
+            'name' => $this->getName(),
+            'url'  => $this->getUrl(),
+            ...$this->getExtensions(),
+        ], fn ($value) => $value !== null);
     }
 }
