@@ -2,9 +2,9 @@
 
 namespace AutoDocumentation\Properties;
 
-use AutoDocumentation\Components\SchemaComponent;
-use AutoDocumentation\Helpers\SchemaHelper;
+
 use AutoDocumentation\Schemas\ArraySchema;
+use AutoDocumentation\Schemas\IntegerSchema;
 use AutoDocumentation\Schemas\Schema;
 
 /**
@@ -15,12 +15,8 @@ use AutoDocumentation\Schemas\Schema;
  */
 class ArrayProperty extends Property
 {
-    public static function make(string $name, SchemaComponent|Schema $items): static
+    public static function make(string $name, Schema $items): static
     {
-        if ($items instanceof Schema || $items instanceof SchemaComponent) {
-            $items = ArraySchema::make($items);
-        }
-
-        return new static($name, SchemaHelper::prepareSchema($items));
+        return new static($name, ArraySchema::make($items));
     }
 }
