@@ -3,7 +3,10 @@
 namespace AutoDocumentation\Properties;
 
 use AutoDocumentation\Schemas\Schema;
-use AutoDocumentation\Traits\CanBeRequired;
+use AutoDocumentation\Traits\HasName;
+use AutoDocumentation\Traits\HasRequired;
+use AutoDocumentation\Traits\HasSchema;
+use Illuminate\Contracts\Support\Arrayable;
 
 /**
  * @method self title(string $title)
@@ -12,32 +15,30 @@ use AutoDocumentation\Traits\CanBeRequired;
  * @method self default($default)
  * @method self description(string $description)
  * @method self example(string $example)
+ * @method self extension(string $name, mixed $value)
  *
  * @see Schema
  */
-class Property
+class Property implements Arrayable
 {
-    use CanBeRequired;
-
-    protected string $name;
-    protected Schema $schema;
+    use HasName;
+    use HasRequired;
+    use HasSchema;
 
     public function __construct(string $name, Schema $schema)
     {
-        $this->name = $name;
-        $this->schema = $schema;
+        $this->name($name);
+        $this->schema($schema);
     }
 
-    public function resolve(): array
+    public function toArray(): array
     {
-        return [
-            $this->name => $this->schema->resolve(),
-        ];
+        return $this->schema->toArray();
     }
 
     public function __call(string $method, array $arguments): static
     {
-        $this->schema->{$method}(...$arguments);
+        $this->getSchema()->{$method}(...$arguments);
 
         return $this;
     }
