@@ -4,7 +4,7 @@ namespace AutoDocumentation\Responses;
 
 class SuccessfulResponse extends Response
 {
-    public static function make(int $statusCode = 200): Response
+    public static function make(int $statusCode = 200): static
     {
         return parent::make($statusCode)->description('Successful operation');
     }
