@@ -2,42 +2,35 @@
 
 namespace AutoDocumentation\Schemas;
 
-use AutoDocumentation\Contracts\Resolvable;
 use AutoDocumentation\Enums\Type;
-use AutoDocumentation\Traits\CanBeDeprecated;
+use AutoDocumentation\Traits\HasDeprecated;
 use AutoDocumentation\Traits\HasDescription;
+use AutoDocumentation\Traits\HasEnum;
+use AutoDocumentation\Traits\HasExample;
+use AutoDocumentation\Traits\HasExamples;
+use AutoDocumentation\Traits\HasExtensions;
 use AutoDocumentation\Traits\HasExternalDocs;
+use AutoDocumentation\Traits\HasFormat;
+use AutoDocumentation\Traits\HasTitle;
+use Illuminate\Contracts\Support\Arrayable;
 
-abstract class Schema implements Resolvable
+abstract class Schema implements Arrayable
 {
+    use HasTitle;
     use HasDescription;
-    use CanBeDeprecated;
+    use HasDeprecated;
     use HasExternalDocs;
+    use HasFormat;
+    use HasEnum;
+    use HasExample;
+    use HasExamples;
+    use HasExtensions;
 
     protected Type $type;
-    protected ?string $title = null;
     protected ?string $format = null;
     protected mixed $default = null;
-    protected ?array $enum = null;
 
     protected bool $nullable = false;
-    protected bool $readOnly = false;
-    protected bool $writeOnly = false;
-    protected ?string $example = null;
-
-    public function title(string $title): static
-    {
-        $this->title = $title;
-
-        return $this;
-    }
-
-    public function format(string $format): static
-    {
-        $this->format = $format;
-
-        return $this;
-    }
 
     public function default($default): static
     {
@@ -46,48 +39,22 @@ abstract class Schema implements Resolvable
         return $this;
     }
 
-    public function enum(array|string $enum): static
+    public function getType(): Type
     {
-        if (is_string($enum)) {
-            $enum = array_map(fn(\BackedEnum $e) => $e->value, $enum::cases());
-        }
-
-        $this->enum = $enum;
-
-        return $this;
+        return $this->type;
     }
 
-    public function example(string $example): static
-    {
-        $this->example = $example;
-
-        return $this;
-    }
-
-    public function resolve(): array
+    public function toArray(): array
     {
         return [
-            'type'        => $this->type->name,
-            'title'       => $this->title,
-            'format'      => $this->format,
-            'default'     => $this->prepareDefault($this->default),
-            'enum'        => $this->enum,
-            'description' => $this->description,
-            ...$this->additionalFields(),
+            'type'        => $this->getType()->value,
+            'title'       => $this->getTitle(),
+            'format'      => $this->getFormat(),
+//            'default'     => $this->prepareDefault($this->default),
+            'enum'        => $this->getEnum(),
+            'description' => $this->getDescription(),
+            'example'     => $this->getExample(),
+            ...$this->getExtensions(),
         ];
-    }
-
-    protected function additionalFields(): array
-    {
-        return [];
-    }
-
-    protected function prepareDefault($default)
-    {
-        if ($default instanceof self) {
-            return $default->resolve();
-        }
-
-        return $default;
     }
 }
