@@ -2,7 +2,7 @@
 
 namespace AutoDocumentation\Tests;
 
-use AutoDocumentation\PackageServiceProvider;
+use AutoDocumentation\AutoDocumentationServiceProvider;
 use Orchestra\Testbench\TestCase;
 
 class PackageTestCase extends TestCase
@@ -10,7 +10,7 @@ class PackageTestCase extends TestCase
     protected function getPackageProviders($app): array
     {
         return [
-            PackageServiceProvider::class,
+            AutoDocumentationServiceProvider::class,
         ];
     }
 }
